refactor(core): extract shared http_builders for method, body, and client config
